+++ i21Database/dbo/Stored Procedures/uspSTstgInsertMixMatchFile.sql
@@ -40,7 +40,6 @@ BEGIN
 					SET @strGeneratedXML = ''
 					SET @intImportFileHeaderId = 0
 					SET @strMessageResult = 'Register ' + @strRegister + ' has no Outbound setup for Promotion Sales List - Mix and Match (' + @strFilePrefix + ')'
-					-- SET @strResult = 'Register ' + @strRegister + ' has no Outbound setup for Send Promotion Sales List File (' + @strFilePrefix + ')'
 
 					RETURN
 			END
@@ -140,7 +139,7 @@ BEGIN
 							ELSE 
 								BEGIN
 									SET @ysnSuccessResult = CAST(0 AS BIT)
-									SET @strMessageResult = 'No result found'
+									SET @strMessageResult = 'No result found to generate Mix/Match - ' + @strFilePrefix + ' Outbound file'
 								END
 					END
 			END
@@ -228,15 +227,22 @@ BEGIN
 				AND ST.intStoreId = @intStoreId 
 				AND PSL.strPromoType = 'M' -- <--- 'M' = Mix and Match
 				-- AND PSL.intPromoSalesId BETWEEN @BeginningMixMatchId AND @EndingMixMatchId
+				
 
-				--SELECT @intImportFileHeaderId = intImportFileHeaderId FROM dbo.tblSMImportFileHeader 
-				--Where strLayoutTitle = 'Pricebook Mix Match' AND strFileType = 'XML'
-	
-				--Generate XML for the pricebook data availavle in staging table
-				Exec dbo.uspSMGenerateDynamicXML @intImportFileHeaderId, 'tblSTstgMixMatchFile~intMixMatchFile > 0', 0, @strGeneratedXML OUTPUT
 
-				--Once XML is generated delete the data from pricebook  staging table.
-				DELETE FROM tblSTstgMixMatchFile	
+				IF EXISTS(SELECT StoreLocationID FROM tblSTstgMixMatchFile)
+					BEGIN
+							-- Generate XML for the pricebook data availavle in staging table
+							EXEC dbo.uspSMGenerateDynamicXML @intImportFileHeaderId, 'tblSTstgMixMatchFile~intMixMatchFile > 0', 0, @strGeneratedXML OUTPUT
+
+							-- Once XML is generated delete the data from pricebook  staging table.
+							DELETE FROM tblSTstgMixMatchFile
+					END
+				ELSE 
+					BEGIN
+							SET @ysnSuccessResult = CAST(0 AS BIT)
+							SET @strMessageResult = 'No result found to generate Mix/Match - ' + @strFilePrefix + ' Outbound file'
+					END
 			END
 
 	END TRY
